+++ src/test/java/at/or/reder/frodo/modbus/service/MarketPriceSchedulerServiceTest.java
@@ -17,41 +17,41 @@ class MarketPriceSchedulerServiceTest {
   @Test
   void conversion_typicalPositivePrice() {
     // 100 EUR/MWh → 10 ct/kWh
-    assertEquals(10.0, MarketPriceSchedulerService.eurMwhToCtKwh(100.0), DELTA);
+    assertEquals(10.0, MarketPriceSchedulerService.eurMwhToCtKwh(100.0).doubleValue(), DELTA);
   }
 
   @Test
   void conversion_zero() {
-    assertEquals(0.0, MarketPriceSchedulerService.eurMwhToCtKwh(0.0), DELTA);
+    assertEquals(0.0, MarketPriceSchedulerService.eurMwhToCtKwh(0.0).doubleValue(), DELTA);
   }
 
   @Test
   void conversion_negativePrice_remainsNegative() {
     // Negative EUR/MWh → negative ct/kWh (triggers export block)
-    assertEquals(-5.0, MarketPriceSchedulerService.eurMwhToCtKwh(-50.0), DELTA);
+    assertEquals(-5.0, MarketPriceSchedulerService.eurMwhToCtKwh(-50.0).doubleValue(), DELTA);
   }
 
   @Test
   void conversion_smallNegativePrice() {
     // -1 EUR/MWh → -0.1 ct/kWh  (still negative → still blocked)
-    assertEquals(-0.1, MarketPriceSchedulerService.eurMwhToCtKwh(-1.0), DELTA);
+    assertEquals(-0.1, MarketPriceSchedulerService.eurMwhToCtKwh(-1.0).doubleValue(), DELTA);
   }
 
   @Test
   void conversion_largePositivePrice() {
     // 500 EUR/MWh → 50 ct/kWh
-    assertEquals(50.0, MarketPriceSchedulerService.eurMwhToCtKwh(500.0), DELTA);
+    assertEquals(50.0, MarketPriceSchedulerService.eurMwhToCtKwh(500.0).doubleValue(), DELTA);
   }
 
   @Test
   void conversion_fractionalInput() {
     // 13.7 EUR/MWh → 1.37 ct/kWh
-    assertEquals(1.37, MarketPriceSchedulerService.eurMwhToCtKwh(13.7), DELTA);
+    assertEquals(1.37, MarketPriceSchedulerService.eurMwhToCtKwh(13.7).doubleValue(), DELTA);
   }
 
   @Test
   void conversion_scaleFactor_isExactlyOneTenth() {
     // The scale factor must be exactly 1/10 — not 0.099 or 0.101
-    assertEquals(MarketPriceSchedulerService.eurMwhToCtKwh(10.0), 1.0, DELTA);
+    assertEquals(MarketPriceSchedulerService.eurMwhToCtKwh(10.0).doubleValue(), 1.0, DELTA);
   }
 }
